Give a name to the price list result
Using let we can define a locally scoped name for the results of taking 10 items
from generate-prices.

Using a def function, we can define a name that can be used throughout the
namespace to represent the generated price list.  This pricelist name points to
the infinite collection of prices.  This name is lazily evaluated so we can
decide later when the name is actually used as to how many prices we want to
work with.