Creating prices in a map, along with time data
Now we generate prices into a map.

First we create a map collection of :price and value.  This is done by mapping
an anonymous function over the pricelist and placing the argument (the next
value from pricelist) into the map as a value along with the :price key.

Then we create a map collection with price and time.  This is done by mapping a
function over two other anonymous functions.

The first anonymous function, taken as the times argument to the parent map,
creates a map with a :time key and an incremented integer, starting from zero.

The second anonymous function is the same as the initial exaple where we map
over pricelist and create a :price key with the next value from pricelist.

The parent map function is within a take function, so that the child map
functions know how many values to extract.